requirements.txt
matplotlib==3.8.3
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.0
Pillow==9.5.0
Pillow==10.3.0
scikit_learn==1.5.0
scipy==1.13.1
torch==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
torchviz==0.0.2
tqdm==4.66.2
